[#78] run_pending_job를 service 계층으로 재작성
