+++ src/actions/mod.rs
@@ -1,2 +1,1 @@
-pub mod run_pending_job;
 pub mod track_runnng_job;
