+++ src/background/runner.rs
@@ -1,5 +1,4 @@
 use crate::{
-    actions::run_pending_job::run_pending_job,
     context::SharedContext,
     domain::job::{
         dao::{ListJobsParams, PatchJobParams},
@@ -36,7 +35,7 @@ pub async fn start_runner_loop(context: SharedContext) {
             }
 
             for pending_job in pending_jobs {
-                if let Err(error) = run_pending_job(context.clone(), &pending_job).await {
+                if let Err(error) = context.job_service.run_pending_job(&pending_job).await {
                     println!("Error processing job: {:?}", error);
 
                     let patch_result = context
