+++ src/domain/job.rs
@@ -18,4 +18,5 @@ pub trait JobRepository {
 pub trait JobService {
     async fn submit_job(&self, params: SubmitJobRequest) -> anyhow::Result<i64>;
     async fn stop_job(&self, params: StopJobRequest) -> anyhow::Result<()>;
+    async fn run_pending_job(&self, pending_job: &entities::job::Model) -> anyhow::Result<()>;
 }
